requirements.txt
safe_pysha3==1.0.4
